Fix sign conversions in menus, options and other UI code
Menu enable masks are uint32_t, load-dialog dates int64_t, profile
positions, palette luminance and keyframe offsets signed, and TD's mono
screen characters unsigned char like RA's. RA's unit tracker totals are
int32_t, so a 64-bit build now sends the 4-byte values the stats packet
declares. Library lengths go through base::ToSize and base::ToSigned, and
pixel, palette and shape data stay unsigned and bit-identical.

+++ src/ra/ini.h
@@ -148,7 +148,9 @@ class INIClass {
     INIEntry& operator=(const INIEntry&) = delete;
     INIEntry(INIEntry&&) = delete;
     INIEntry& operator=(INIEntry&&) = delete;
-    [[nodiscard]] int Index_ID() const { return CrcEngine::Compute(Entry); }
+    [[nodiscard]] int Index_ID() const {
+      return static_cast<int>(CrcEngine::Compute(Entry));
+    }
 
     std::string Entry;
     std::string Value;
@@ -166,7 +168,9 @@ class INIClass {
     INISection(INISection&&) = delete;
     INISection& operator=(INISection&&) = delete;
     INIEntry* Find_Entry(const char* entry) const;
-    [[nodiscard]] int Index_ID() const { return CrcEngine::Compute(Section); }
+    [[nodiscard]] int Index_ID() const {
+      return static_cast<int>(CrcEngine::Compute(Section));
+    }
 
     std::string Section;
     List<INIEntry> EntryList;
